No real bugs this time, only super ugly code.

* clang-implicit-conversion-sanitizer-compound-assignment-operator:
  NefDecoder::decodeMetaDataInternal(): decryption is modulo-2^n math
  DeflateDecompressor: decodeFPDeltaRow() works in modulo-2^n arith.
  HuffmanTableLUT::setup(): and one more place where the signedness changes.

https://reviews.llvm.org/D53949

+++ src/librawspeed/decompressors/HuffmanTableLUT.h
@@ -156,8 +156,9 @@ public:
 
           if (diff_l) {
             uint32 diff = (c >> (LookupDepth - code_l - diff_l)) & ((1 << diff_l) - 1);
-            decodeLookup[c] |= static_cast<uint32>(signExtended(diff, diff_l))
-                               << PayloadShift;
+            decodeLookup[c] |= static_cast<int32>(
+                static_cast<uint32>(signExtended(diff, diff_l))
+                << PayloadShift);
           }
         }
       }
